requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
openpyxl>=3.1.0

# Geo
geopandas>=0.14.0
fiona>=1.9.0
rasterio>=1.3.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.43.0
PyWavelets>=1.4

# DL / CV
torch>=2.0.0
segmentation-models-pytorch>=0.3.3